pyproject.toml
[tool.poetry]
name = "nonebot-plugin-treehelp"
version = "0.0.1"
description = "适用于 Nonebot2 的树形帮助插件"
authors = ["hemengyang <hmy0119@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/he0119/nonebot-plugin-treehelp"
repository = "https://github.com/he0119/nonebot-plugin-treehelp"
documentation = "https://github.com/he0119/nonebot-plugin-treehelp#readme"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.5"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pre-commit = "*"
nonebug = "^0.2.2"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.0.2"
pytest-mock = "^3.10.0"
pytest-asyncio = "^0.20.2"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
